feat: implemented serial control in theory
