+++ starr_gazer_hardware_system/pyproject.toml
@@ -6,5 +6,9 @@ readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
     "jetson-gpio>=2.1.9",
+    "numpy>=2.3.3",
+    "opencv-python>=4.11.0.86",
+    "pygame>=2.6.1",
     "pyserial>=3.5",
+    "ultralytics>=8.3.204",
 ]
